Team page
This page has a contact form on it so visitors can send in questions or comments about topics having to do with substance abuse. The form will also be linked to a post page where visitors can post comments or questions on a blog. I need to fix the search function to  search google as well as the pages on www.computerlove69.com the main page to this team page.